+++ tailwind.config.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 module.exports = {
     purge: ['./pages/**/*.js', './components/**/*.js'],
-    darkMode: false, // or 'media' or 'class'
+    darkMode: 'class', // or 'media' or false
     theme: {
         extend: {},
     },